refactor(app): ModuleInfo to molecules
and useRunHasStarted to resources, which is most of the diff
