+++ app/src/organisms/CalibrationTaskList/__tests__/CalibrationTaskList.test.tsx
@@ -18,9 +18,10 @@ import {
   expectedIncompleteRightMountTaskList,
   expectedIncompleteLeftMountTaskList,
 } from '../../Devices/hooks/__fixtures__/taskListFixtures'
-import { useCalibrationTaskList, useRunHasStarted } from '../../Devices/hooks'
+import { useCalibrationTaskList } from '../../Devices/hooks'
 import { useAttachedPipettes } from '/app/resources/instruments'
 import { mockLeftProtoPipette } from '/app/redux/pipettes/__fixtures__'
+import { useRunHasStarted } from '/app/resources/runs'
 
 vi.mock('../../Devices/hooks')
 vi.mock('/app/resources/runs')
