+++ app/src/organisms/Devices/ProtocolRun/SetupModuleAndDeck/__tests__/SetupModulesMap.test.tsx
@@ -15,8 +15,8 @@ import {
   mockMagneticModule as mockMagneticModuleFixture,
 } from '/app/redux/modules/__fixtures__/index'
 import { useMostRecentCompletedAnalysis } from '/app/resources/runs'
-import { getAttachedProtocolModuleMatches } from '../../../../ODD/ProtocolSetup/ProtocolSetupModulesAndDeck/utils'
-import { ModuleInfo } from '../../../ModuleInfo'
+import { getAttachedProtocolModuleMatches } from '/app/organisms/ODD/ProtocolSetup/ProtocolSetupModulesAndDeck/utils'
+import { ModuleInfo } from '/app/molecules/ModuleInfo'
 import { SetupModulesMap } from '../SetupModulesMap'
 
 import type {
@@ -44,8 +44,8 @@ vi.mock('@opentrons/shared-data', async importOriginal => {
   }
 })
 vi.mock('/app/resources/runs/useMostRecentCompletedAnalysis')
-vi.mock('../../../../ODD/ProtocolSetup/ProtocolSetupModulesAndDeck/utils')
-vi.mock('../../../ModuleInfo')
+vi.mock('/app/organisms/ODD/ProtocolSetup/ProtocolSetupModulesAndDeck/utils')
+vi.mock('/app/molecules/ModuleInfo')
 vi.mock('/app/resources/modules')
 
 const render = (props: React.ComponentProps<typeof SetupModulesMap>) => {
