+++ app/src/organisms/Devices/ProtocolRun/SetupModuleAndDeck/index.tsx
@@ -22,8 +22,8 @@ import {
   getRequiredDeckConfig,
 } from '/app/resources/deck_configuration/utils'
 import { useRobotType } from '/app/redux-resources/robots'
+import { useRunHasStarted } from '/app/resources/runs'
 import {
-  useRunHasStarted,
   useUnmatchedModulesForProtocol,
   useModuleCalibrationStatus,
 } from '../../hooks'
