+++ app/src/organisms/Devices/ProtocolRun/SetupRobotCalibration.tsx
@@ -16,11 +16,10 @@ import {
   ANALYTICS_PROCEED_TO_MODULE_SETUP_STEP,
   ANALYTICS_PROCEED_TO_LABWARE_SETUP_STEP,
 } from '/app/redux/analytics'
-import { useRunHasStarted } from '../hooks'
 import { SetupDeckCalibration } from './SetupDeckCalibration'
 import { SetupInstrumentCalibration } from './SetupInstrumentCalibration'
 import { SetupTipLengthCalibration } from './SetupTipLengthCalibration'
-import { useRunStatus } from '/app/resources/runs'
+import { useRunStatus, useRunHasStarted } from '/app/resources/runs'
 import { RUN_STATUS_STOPPED } from '@opentrons/api-client'
 import { useIsFlex } from '/app/redux-resources/robots'
