+++ app/src/organisms/Devices/ProtocolRun/__tests__/SetupTipLengthCalibrationButton.test.tsx
@@ -9,7 +9,8 @@ import { renderWithProviders } from '/app/__testing-utils__'
 import { i18n } from '/app/i18n'
 import { mockDeckCalData } from '/app/redux/calibration/__fixtures__'
 import { mockTipLengthCalLauncher } from '../../hooks/__fixtures__/taskListFixtures'
-import { useDeckCalibrationData, useRunHasStarted } from '../../hooks'
+import { useDeckCalibrationData } from '../../hooks'
+import { useRunHasStarted } from '/app/resources/runs'
 import { useDashboardCalibrateTipLength } from '/app/pages/Desktop/Devices/CalibrationDashboard/hooks/useDashboardCalibrateTipLength'
 import { SetupTipLengthCalibrationButton } from '../SetupTipLengthCalibrationButton'
 
@@ -23,6 +24,7 @@ vi.mock(
 vi.mock('/app/redux/config')
 vi.mock('/app/redux/sessions/selectors')
 vi.mock('../../hooks')
+vi.mock('/app/resources/runs')
 
 const ROBOT_NAME = 'otie'
 const RUN_ID = '1'
