+++ app/src/organisms/Devices/hooks/index.ts
@@ -12,7 +12,6 @@ export * from './useModuleCalibrationStatus'
 export * from './usePipetteOffsetCalibrations'
 export * from './usePipetteOffsetCalibration'
 export * from './useRunCreatedAtTimestamp'
-export * from './useRunHasStarted'
 export * from './useTipLengthCalibrations'
 export * from './useUnmatchedModulesForProtocol'
 export * from './useProtocolAnalysisErrors'
