+++ app/src/organisms/ODD/ProtocolSetup/ProtocolSetupModulesAndDeck/__tests__/ModulesAndDeckMapView.test.tsx
@@ -17,7 +17,7 @@ vi.mock('@opentrons/api-client')
 vi.mock('@opentrons/shared-data/js/helpers/getSimplestFlexDeckConfig')
 vi.mock('/app/redux/config')
 vi.mock('/app/resources/deck_configuration/utils')
-vi.mock('../../../../Devices/ModuleInfo')
+vi.mock('/app/molecules/ModuleInfo')
 vi.mock('/app/transformations/analysis')
 
 const mockRunId = 'mockRunId'
