+++ app/src/resources/runs/index.ts
@@ -20,3 +20,4 @@ export * from './constants'
 export * from './useMostRecentCompletedAnalysis'
 export * from './useRunPipetteInfoByMount'
 export * from './useRunCalibrationStatus'
+export * from './useRunHasStarted'
